steel-language-server: disable markdown feature and fix doc extraction
this fixes two seperate, but slightly related issues:

1. the steel-language-server depended on `steel-core` with the `markdown`
feature enabled, causing the builtin docs to be rendered, instead of
being passed along as the raw markdown.

2. when collecting doc comment lines, steel previously trimmed all
whitespace at the start. this caused code blocks with indentation to
break, as the indentation was stripped as well. additionally steel did
not attempt to trim the trailing newlines, and then later joined all the
comment lines with a "\n", resulting in the newlines getting duplicated.