implement `inexact` from r7rs (#381)
* deduplicate inexact->exact impl, implement inexact procedure

* swap inexact->exact and exact impls (and the same for inexact)

inexact->exact and exact->inexact was removed in r7rs and replaced with
just exact and inexact respectively. removing them is not a very smart
idea as it just breaks code without any benefit, but moving the actual
impl into the new "canonical" functions just makes more sense.

* give an error when trying to convert nan or inf to an exact value

it would previously just return `#false`, due to the impl IntoSteelval
for Option, which i don't think was intended. this changes it to return
an error.

* run cargo xtask docgen