return a unicode replacement character instead of erroring in read-char
this seems to be what most other scheme interpreters, at least the ones,
that i have tested this with (racket, mit-scheme, guile, tinyscheme), do.
none that i could find return an error, and to.

both r5rs and r7rs are pretty vague about it, only specifying, that
`read-char` "returns the next character," but given, that it specifically
isn't mentioned, that it is "an error," which both documents frequently
do, i think it should not return an error.