don't flush on write
the spec doesn't say anything about flushing, and there is a procedure
specifically to flushing, so i think it makes a lot more sense to not
always flush for performance reasons, as always flushing kind of defeats
the point of buffered-writing.

additionally, when `strace`-ing guile, it doesn't flush on write either.