always print fract part of floats
before:

```
λ > (displayln 1.0)
1
```

after:

```
λ > (displayln 1.0)
1.0
```

because i changed from `Display` to `Debug` printing the floats this also
affects some more things, specifically this also changes the printing of
large floats to use scientific notation.
this should not affect the precision, as it only does scientific notation
if the rest of the numbers are all 0, and is actually what other scheme
dialects do as well.

before:

```
λ > (displayln 900719925474099113909139183.0)
900719925474099000000000000
```

after:

```
λ > (displayln 900719925474099113909139183.0)
9.00719925474099e26
```

note, that after the 99 it was followed only by 0s, so we didn't actually
lose any precision.

racket:

```
> (displayln 900719925474099113909139183.0)
9.00719925474099e+26
```