vector-set!/swap!: fix docs and crash (#496)
* fix docs for vector-swap!

the previous docs for vector-swap! were essentially just copied from
vector-set!, so they weren't very helpful.

* fix bounds checks in vector-set!, vector-swap!

previously it was relatively easy to accidentally crash steel due to
using a > instead of the more appropriate >= in a bounds check. in
addition, the bounds check for the second index in vector-swap! was
forgotten.

```
$ steel
λ > (define v (mutable-vector 1 2 3))
λ > (vector-set! v 3 4)
thread 'main' panicked at crates/steel-core/src/primitives/vectors.rs:950:10:
index out of bounds: the len is 3 but the index is 3
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```