remove unused bool return from read_bytes_amt
i wanted to add some (unrelated to this pr) features, and to do that i
had to refactor the read_bytes_amt fn, where i stumbled upon the bool
return parameter, which i didn't know what it was supposed to indicate.
so i checked where it was used to figure it out, und it turns out it's
not used at all, so i can just remove it.