improve port `write` (#501)
* don't flush on write

the spec doesn't say anything about flushing, and there is a procedure
specifically to flushing, so i think it makes a lot more sense to not
always flush for performance reasons, as always flushing kind of defeats
the point of buffered-writing.

additionally, when `strace`-ing guile, it doesn't flush on write either.

* remove now-redundant write_and_flush macro

since there is now only one call, as the flush call is no longer there,
i can actually just inline it and it's actually shorter.

* use write_all over just write

calls to write can just decide not to write the entire input to the
buffer, so use `write_all`, which is a lot more sensible.