flush stdout in repl after user prompt
when i changed the buffering i said that nothing relies on the flushing
behaviour of write. turns out, this is incorrect. i still stand by the
fact that nothing *should* rely on that behaviour, but the repl actually
does.

before:

```
λ > (display 'symbol)
λ > (write "string")
λ > (newline)
symbol"string"
λ > (display (list 1 2 3))
λ >
(1 2 3)CTRL-D
```

after:

```
λ > (display 'symbol)
symbol
λ > (write "string")
"string"
λ > (newline)

λ > (display (list 1 2 3))
(1 2 3)
λ >
CTRL-D
```

this was actually observable before my change as well, when using
`(write-char)` in the repl, but it is now significantly more obvious, as
it now affects all write calls without a newline.

```
λ > (write-char #\w)
λ > (write-char #\h)
λ > (write-char #\y)
λ > (newline)
why
```