implement the `(peek-char)` procedure (#516)
* add Peekable wrapper struct around input ports

currently this does not change anything in functionality, but providing
a wrapper struct around input ports, which can then in the future be
extended to hold a buffer. currently this only holds an Option to
provide the already existing byte-peeking behaviour.

it is a little unfortunate, that i still need the port_read_str_fn!
macro, as `Stdin` is not `: BufRead`, but still provides `read_line` and
`read_to_string` and, due to orphan rules, i cannot implement a function
for both `Peekable<R: BufRead>` and `Peekable<Stdin>`.

* add a 4-wide buffer to Peekable in preparation for peek-char

* implement peek-char

* add some port tests for (peek-byte), (peek-char), (read-char) etc