implement and fix case-folding / case insensitive procedures (#523)
* add some string equality tests

currently most of these are not implemented or will fail, but having
tests, that are then commented out makes it a little more obvious where
it is currently incorrect.

as a drive-by, this also includes the docs i forgot to generate in my
last pr.

* implement char-ci=? via icu_casemap

* implement the rest of the case-insensitive char comparisons

* implement char-foldcase, fix char-upcase and char-downcase

previously char-upcase and char-downcase were applying full case
mappings and then taking the first char instead of simple case mappings,
resulting in unexpected behavior such as (char-upcase #\ß) ;; => #\S

* rename procedures to string-upcase and string-downcase

... and then add the previous names as aliases.

this is to align the procedure names with r7rs without breaking backwards
compatibility.

* implement string-foldcase

* do full case-folding in string equality procedures

* add examples to the string/char equality/case mapping procedures

* add examples to char comparison procedures