move quotient and remainder code into the truncate-<op> definitions
while r7rs defines the quotient and remainder proceduers, it says, that
those are only "provided for backward compatibility" with r5rs, so i
think it makes sense to move the code into the canonical procedures.

i have decided against defining those as an actual alias for their
canonical versions, as i think those are the procedures people are more
likely to search for, making it valuable to find documentation
immediately.