make symbol=? work with uninterned symbols
the scheme spec says, that `eqv?` should return #t for two symbols, if
they are equal according to `symbol=?`. previously, `symbol=?` would
return `#f`, if the symbols were `string=?`, but one (or both) of the
symbols were uninterned. this was in conflict with `eqv?`, which would
return `#t` regardless of if the arguments were interned or not.