rename default-current-<direction>-port to just default-<direction>-port
those procedures were previously also used in the read-port etc
procedures, but since those have been moved to the steel side to fix
parameterizing, these are no longer needed as the rust-side private
builtins can just take a port unconditionally.

in my opinion calling them something with "current" is a little
misleading, as these are just constructors for the default error ports,
as they are not affected by parameterizing, which the
current-<direction>-port should be. additionally it the functions better
up with the private builtin wrapper procedures (their only call site),
as those are just called `#%default-<direction>-port`.