fix and make fs docs more in line with other docs
most of the rest of the docs end the each "paragraph" with a full stop,
while these didn't.

additionally the lines with the parameters specified their type in
paranthesis, like `path : (string?)`, while the docs for all other
procedures specify it like `path : string?`.

and as a last thing, almost all of the procedures described the path
argument as "The path to check", even if they weren't checking anything
and were just retrieving something, so fix that too.