misc error and docs stuff (#547)
* fix usage docs of (close-output-port), (close-input-port)

* fix and make fs docs more in line with other docs

most of the rest of the docs end the each "paragraph" with a full stop,
while these didn't.

additionally the lines with the parameters specified their type in
paranthesis, like `path : (string?)`, while the docs for all other
procedures specify it like `path : string?`.

and as a last thing, almost all of the procedures described the path
argument as "The path to check", even if they weren't checking anything
and were just retrieving something, so fix that too.

* fix example for list-ref

* slightly improve real-part and imag-part error messages

actually print, what was passed to the procedure in the TypeMismatch
errors in `real-part` and `imag-part`. also changes the tense of the
error to be more in line with the other errors, which are usually in
present tense instead of past.

before:

```
λ > (real-part #f)
error[E03]: TypeMismatch
  ┌─ :1:2
  │
1 │ (real-part #f)
  │  ^^^^^^^^^ real-part expected number
```

after:

```
λ > (real-part #f)
error[E03]: TypeMismatch
  ┌─ :1:2
  │
1 │ (real-part #f)
  │  ^^^^^^^^^ real-part expects a number, found #false
```

* correctly use steelerr! macro format string

the `steelerr!` and related macros actually only interpret the argument
as a format string, if there are multiple arguments, so it doesn't work
if you use inlined format args. this fixes a few usages of it, so that
the errors actually make sense.

before:

```
λ > (angle #f)
error[E03]: TypeMismatch
  ┌─ :1:2
  │
1 │ (angle #f)
  │  ^^^^^ angle expects a number, found {number}

λ > (magnitude #f)
error[E03]: TypeMismatch
  ┌─ :1:2
  │
1 │ (magnitude #f)
  │  ^^^^^^^^^ magnitude expects a number, found {number}

λ > (exact-integer-sqrt #f)
error[E03]: TypeMismatch
  ┌─ :1:2
  │
1 │ (exact-integer-sqrt #f)
  │  ^^^^^^^^^^^^^^^^^^ exact-integer-sqrt expects a non-negative integer but found {number}
````

after:

```
λ > (angle #f)
error[E03]: TypeMismatch
  ┌─ :1:2
  │
1 │ (angle #f)
  │  ^^^^^ angle expects a number, found #false

λ > (magnitude #f)
error[E03]: TypeMismatch
  ┌─ :1:2
  │
1 │ (magnitude #f)
  │  ^^^^^^^^^ magnitude expects a number, found #false

λ > (exact-integer-sqrt #f)
error[E03]: TypeMismatch
  ┌─ :1:2
  │
1 │ (exact-integer-sqrt #f)
  │  ^^^^^^^^^^^^^^^^^^ exact-integer-sqrt expects a non-negative integer but found #false
```

* correct error from TypeMismatch to ArityMismatch

* don't print "Parse" twice on a ParseError

before:

```
λ > (/ 1 2]
error[E09]: Parse
  ┌─ :1:7
  │
1 │ (/ 1 2]
  │       ^ Parse: Mismatched parenthesis, expected ")"
```

after:

```
λ > (/ 1 2]
error[E09]: Parse
  ┌─ :1:7
  │
1 │ (/ 1 2]
  │       ^ Mismatched parenthesis, expected ")"
```

* run cargo xtask docgen, as i forgot that in the quotient/remainder pr