misc cleanup and docs in stdlib.scm (#558)
* raise an error in (drop) when called with negative n

this matches the behaviour of `(take)`

before:

```
λ > (take '(1 2 3 4) -1)
error[E11]: Generic
  ┌─ :1:2
  │
1 │ (take '(1 2 3 4) -1)
  │  ^^^^ take expects a positive integer

λ > (drop '(1 2 3 4) -1)
error[E11]: Generic
    ┌─ :652:37
    │
652 │     (if (zero? x) l (loop (sub1 x) (cdr l))))
    │                                     ^^^ cdr expects a non empty list
```

after:

```
λ > (drop '(1 2 3 4) -1)
error[E11]: Generic
    ┌─ :671:8
    │
671 │       (error 'drop "expects a positive number")
    │        ^^^^^  drop expects a positive number
```

also add some documentation.

* cleanup and sort foldl foldr unfold etc

- use shorthand procedure definition more consistently
- put foldl directly next to foldr
- just define reduce and fold as an alias instead
- rename the pred in unfold to stop?

* use cond in assoc procedures, actually use eqv? in assv

i know that the `eqv?` and `eq?` procedures actually do exactly the same
in steel, but it is still a little strange, when `assv` does not use the
`eqv?` procedure.

this also, as usual, adds some docs.

* vastly simplify for-each

if the list is null? is already checked at the start.

and also add docs.

* remove all the duplicate for-each definitions

for-each is in the stdlib, so everything should be able to just call
that function instead of defining one for themselves.

* make (flatten) work with pairs as well

before:

```
λ > (flatten '(a (b . c)))
=> '(a (b . c))
```

after:

```
λ > (flatten '(a (b . c)))
=> '(a b c)
```

also add docs

* cleanup and document the member procedures

- remove the commented out member and memq procedures in stdlib.scm

not a lot of use keeping them around, as creating them again is only a
matter of copying the memv procedure and changing eqv? to equal? / eq?.

- use shorthand procedure define for memv
- explicitly (provide memv)
- add docs to member, memv and memq