allow giving more args to `apply` (#582)
* allow giving more args to `apply`

the scheme spec says, that apply can take more arguments, that are then
prepended to the given list, before that is then given to the procedure.

this is to say, that `(apply proc item lst)` is equivalent to
`(apply proc (cons item lst))` and, more generally, equivalent to
`(apply proc (append (list item ...) lst))`.

* adjust docs for apply procedure

* add some tests for multi-argument apply